Basics of Jpa

+++ src/main/java/com/arun/dao/CustomerDao.java
@@ -3,5 +3,5 @@ package com.arun.dao;
 import com.arun.model.Customer;
 
 public interface CustomerDao {
-    Customer getCustomer(Integer id);
+    Customer getCustomer(Long id);
 }
